Arm and GripActuator: common code in Actuator.py
